+++ src/apps/across/across.definition.ts
@@ -24,16 +24,12 @@ export const ACROSS_DEFINITION = appDefinition({
   links: {
     learn: 'https://docs.across.to/bridge/',
     github: 'https://github.com/across-protocol',
-    twitter: '',
-    telegram: '',
+    twitter: 'https://twitter.com/AcrossProtocol',
     discord: 'https://discord.gg/across',
-    medium: '',
+    medium: 'https://medium.com/across-protocol',
   },
   supportedNetworks: {
     [Network.ETHEREUM_MAINNET]: [AppAction.VIEW],
-    [Network.OPTIMISM_MAINNET]: [AppAction.VIEW],
-    [Network.POLYGON_MAINNET]: [AppAction.VIEW],
-    [Network.ARBITRUM_MAINNET]: [AppAction.VIEW],
   },
   primaryColor: '#fff',
 });
